+++ common/src/main/java/com/chefmooon/colourfulclocks/common/core/PendulumTypes.java
@@ -1,38 +1,46 @@
 package com.chefmooon.colourfulclocks.common.core;
 
+import com.chefmooon.colourfulclocks.common.registry.ColourfulClocksItems;
+import com.chefmooon.colourfulclocks.common.registry.ColourfulClocksSounds;
+import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.sounds.SoundEvent;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 
 public enum PendulumTypes {
 
-    IRON(0, "iron", 2.0F, 1.0F, Items.IRON_INGOT),
-    COPPER(1, "copper", 2.0F, 0.9F, Items.COPPER_INGOT),
-    EXPOSED_COPPER(2, "exposed_copper", 2.0F, 0.8F, ItemStack.EMPTY.getItem()),
-    WEATHERED_COPPER(3, "weathered_copper", 2.0F, 0.7F, ItemStack.EMPTY.getItem()),
-    OXIDIZED_COPPER(4, "oxidized_copper", 2.0F, 0.6F, ItemStack.EMPTY.getItem()),
-    WAXED_COPPER(5, "copper", 2.0F, 0.9F, Items.COPPER_INGOT),
-    WAXED_EXPOSED_COPPER(6, "exposed_copper", 2.0F, 0.8F, ItemStack.EMPTY.getItem()),
-    WAXED_WEATHERED_COPPER(7, "weathered_copper", 2.0F, 0.7F, ItemStack.EMPTY.getItem()),
-    WAXED_OXIDIZED_COPPER(8, "oxidized_copper", 2.0F, 0.6F, ItemStack.EMPTY.getItem()),
-    GOLD(5, "gold", 1.5F, 0.5F, Items.GOLD_INGOT),
-    DIAMOND(6, "diamond", 1.0F, 0.4F, Items.DIAMOND),
-    NETHERITE(7, "netherite", 0.5F, 0.3F, Items.NETHERITE_INGOT)
+    IRON(0, BuiltInRegistries.ITEM.get(ColourfulClocksItems.IRON_PENDULUM), "iron", 2.0F, ColourfulClocksSounds.ITEM_IRON_PENDULUM_CHIME.get(), 1.0F, Items.IRON_INGOT),
+    COPPER(1, BuiltInRegistries.ITEM.get(ColourfulClocksItems.COPPER_PENDULUM), "copper", 2.0F, ColourfulClocksSounds.ITEM_COPPER_PENDULUM_CHIME.get(), 0.9F, Items.COPPER_INGOT),
+    EXPOSED_COPPER(2, BuiltInRegistries.ITEM.get(ColourfulClocksItems.EXPOSED_COPPER_PENDULUM), "exposed_copper", 2.0F, ColourfulClocksSounds.ITEM_EXPOSED_COPPER_PENDULUM_CHIME.get(), 0.8F, ItemStack.EMPTY.getItem()),
+    WEATHERED_COPPER(3, BuiltInRegistries.ITEM.get(ColourfulClocksItems.WEATHERED_COPPER_PENDULUM), "weathered_copper", 2.0F, ColourfulClocksSounds.ITEM_WEATHERED_COPPER_PENDULUM_CHIME.get(), 0.7F, ItemStack.EMPTY.getItem()),
+    OXIDIZED_COPPER(4, BuiltInRegistries.ITEM.get(ColourfulClocksItems.OXIDIZED_COPPER_PENDULUM), "oxidized_copper", 2.0F, ColourfulClocksSounds.ITEM_OXIDIZED_COPPER_PENDULUM_CHIME.get(), 0.6F, ItemStack.EMPTY.getItem()),
+    WAXED_COPPER(5, BuiltInRegistries.ITEM.get(ColourfulClocksItems.WAXED_COPPER_PENDULUM), "copper", 2.0F, ColourfulClocksSounds.ITEM_COPPER_PENDULUM_CHIME.get(), 0.9F, Items.COPPER_INGOT),
+    WAXED_EXPOSED_COPPER(6, BuiltInRegistries.ITEM.get(ColourfulClocksItems.WAXED_EXPOSED_COPPER_PENDULUM), "exposed_copper", 2.0F, ColourfulClocksSounds.ITEM_EXPOSED_COPPER_PENDULUM_CHIME.get(), 0.8F, ItemStack.EMPTY.getItem()),
+    WAXED_WEATHERED_COPPER(7, BuiltInRegistries.ITEM.get(ColourfulClocksItems.WAXED_WEATHERED_COPPER_PENDULUM), "weathered_copper", 2.0F, ColourfulClocksSounds.ITEM_WEATHERED_COPPER_PENDULUM_CHIME.get(), 0.7F, ItemStack.EMPTY.getItem()),
+    WAXED_OXIDIZED_COPPER(8, BuiltInRegistries.ITEM.get(ColourfulClocksItems.WAXED_OXIDIZED_COPPER_PENDULUM), "oxidized_copper", 2.0F, ColourfulClocksSounds.ITEM_OXIDIZED_COPPER_PENDULUM_CHIME.get(), 0.6F, ItemStack.EMPTY.getItem()),
+    GOLD(5, BuiltInRegistries.ITEM.get(ColourfulClocksItems.GOLD_PENDULUM), "gold", 1.5F, ColourfulClocksSounds.ITEM_GOLD_PENDULUM_CHIME.get(), 0.5F, Items.GOLD_INGOT),
+    DIAMOND(6, BuiltInRegistries.ITEM.get(ColourfulClocksItems.DIAMOND_PENDULUM), "diamond", 1.0F, ColourfulClocksSounds.ITEM_DIAMOND_PENDULUM_CHIME.get(), 0.4F, Items.DIAMOND),
+    NETHERITE(7, BuiltInRegistries.ITEM.get(ColourfulClocksItems.NETHERITE_PENDULUM), "netherite", 0.5F, ColourfulClocksSounds.ITEM_NETHERITE_PENDULUM_CHIME.get(), 0.3F, Items.NETHERITE_INGOT)
 
     ;
 
 
 
     private final int id;
+    private final Item item;
     private final String name;
     private final float swingSpeedModifier;
+    private final SoundEvent chimeSound;
     private final float pitchModifier;
     private final Item craftingIngredient;
 
-    PendulumTypes(int id, String name, float swingSpeedModifier, float pitchModifier, Item craftingIngredient) {
+    PendulumTypes(int id, Item item, String name, float swingSpeedModifier, SoundEvent chimeSound, float pitchModifier, Item craftingIngredient) {
         this.id = id;
+        this.item = item;
         this.name = name;
         this.swingSpeedModifier = swingSpeedModifier;
+        this.chimeSound = chimeSound;
         this.pitchModifier = pitchModifier;
         this.craftingIngredient = craftingIngredient;
     }
@@ -41,12 +49,20 @@ public enum PendulumTypes {
         return id;
     }
 
+    public Item getItem() {
+        return item;
+    }
+
     public String getName() {
         return name;
     }
 
     public float getSwingSpeedModifier() {
         return swingSpeedModifier;
+    }
+
+    public SoundEvent getChimeSound() {
+        return chimeSound;
     }
 
     public float getPitchModifier() {