+++ changelog.md
@@ -7,3 +7,6 @@
   - Door must now be open to insert or remove a Pendulum (if door is closed on attempt, it will open)
   - Hand now moves properly when interacting with the door
   - Door will now open/close when struck by a wind charge
+- Pendulum Updates
+  - Each pendulum sound is now separate and can be set with a datapack
+  - New sounds will be added soon